Error on setBitmap method
Inside the setBitmap method, the scope for the variable frame ends once the else block finishes. Thus, calling logFaceData would have no purpose without having a Frame. To fix this either declare the variable frame outside of the else block or call the logFaceData() inside of the else block. In this case, I have chosen to declare the variable frame outside of the else block.